created dynlb module for UISS

+++ Project/UISS/UISSSizeValueConverter.m
@@ -60,7 +60,7 @@
         CGSize size = [converted CGSizeValue];
         
         return [NSString stringWithFormat:@"CGSizeMake(%@, %@)",
-                                          [self.floatValueConverter generateCodeForFloatValue:size.width],
+                                          [self.floatValueConverter generateCodeForFloatValue:(CGFloat)size.width],
                                           [self.floatValueConverter generateCodeForFloatValue:size.height]];
     } else {
         return @"CGSizeZero";
